DiSCo_per

Function to implement permutation test for Distributional Synthetic Controls

Usage

DiSCo_per(
  results.periods,
  T0,
  ww = 0,
  peridx = 0,
  evgrid = seq(from = 0, to = 1, length.out = 101),
  graph = TRUE,
  num_cores = 1,
  redo_weights = FALSE,
  weights = NULL,
  qmethod = NULL,
  q_min = 0,
  q_max = 1,
  M
)

Arguments

results.periods

List of period-specific results from DiSCo

T0

Integer indicating first year of treatment as counted from 1 (e.g, if treatment year 2002 was the 5th year in the sample, this parameter should be 5).

ww

Optional vector of weights indicating the relative importance of each time period. If not specified, each time period is weighted equally.

peridx

Optional integer indicating number of permutations. If not specified, by default equal to the number of units in the sample.

evgrid

Optional vector containing an evenly spaced grid on 0,1 on which the quantile function for the control units will be evaulated. By default, a grid of 100 points is used.

graph

Boolean indicating whether to plot graphs

num_cores

Integer, number of cores to use for parallel computation. Set to 1 by default (sequential computation), this can be very slow!

redo_weights

Boolean indicating whether to recompute the weights for the "true" treated unit. Set to FALSE by default.

weights

Optional vector of weights to use for the "true" treated unit. redo_weights has to be set to FALSE for these weights to be used.

qmethod

Quantile method, see DiSCo function.

q_min

Minimum quantile to use for the permutation test. Default is 0.

q_max

Maximum quantile to use for the permutation test. Default is 1.

M

Number of samples from uniform distribution to use for the permutation test. Default is 1000.

y_name

Y axis label of the graph

x_name

X axis label of the graph

Value

List of matrices containing synthetic time path of the outcome variable for the target unit together with the time paths of the control units

Details

This program iterates through all units and computes the optimal weights on the other units for replicating the unit of iteration's outcome variable, assuming that it is the treated unit. See Algorithm 1 in Gunsilius (2023) for more details. The only modification is that we take the ratio of post- and pre-treatment root mean squared Wasserstein distances to calculate the p-value, rather than the level in each period, following @abadie2010synthetic.

References

Gunsilius FF (2023). “Distributional synthetic controls.” Econometrica, 91(3), 1105--1117.